requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
astroid==3.3.10
attrs==25.3.0
audioop-lts==0.2.1
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dill==0.4.0
discord==2.3.2
discord.py==2.5.2
distlib==0.3.9
dotenv==0.9.9
emoji==2.14.1
filelock==3.18.0
flake8==7.3.0
Flask==3.1.1
frozenlist==1.5.0
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
peewee==3.17.9
peewee-migrate==1.13.0
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
propcache==0.3.0
psutil==7.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
tomlkit==0.13.3
tzdata==2025.1
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
yarl==1.18.3
